Fixed a couple of things in mainroads, postgres.py and added basic ddl sql

+++ sql/transforms/lga_road_stats.sql
@@ -14,7 +14,7 @@ SELECT
     COALESCE(NULLIF(TRIM(lg_name), ''), 'Unknown') AS lg_name,
     COALESCE(NULLIF(TRIM(network_type), ''), 'Unknown') AS network_type,
     COUNT(*)::integer AS segment_count,
-    ROUND(SUM(COALESCE(length_m, 0)) / 1000.0, 3) AS total_length_km
+    ROUND((SUM(COALESCE(length_m, 0)) / 1000.0)::numeric, 3) AS total_length_km
 FROM core.road_segment
 GROUP BY
     COALESCE(NULLIF(TRIM(lg_no), ''), 'UNK'),
